pyproject.toml
[project]
name = "Project3"
version = "0.1.0"
description = "A simple Todo application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.4",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.135.1",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pymysql>=1.1.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
]
